+++ tests/test_resources_positive.py
@@ -12,9 +12,9 @@ class TestResourcesPositive:
         with app.app_context():
             db.create_all()
             app.logger.info("Database initialized")
-            helpers.insert_data_into_resources_table(Resources)
+            helpers.insert_test_data_into_db_table(Resources)
 
-    @pytest.fixture(scope='module')
+    @pytest.fixture(scope="module")
     def client(self):
         testing_client = app.test_client()
         ctx = app.app_context()
@@ -56,44 +56,56 @@ class TestResourcesPositive:
     def test_get_resource_by_id(self, client):
         """ Get resource by its ID. """
         response = client.get("/resources?id=1")
-        exp_json = [{'active': True, 'created_at': '2018-10-10 10:10:10', 'id': 1, 'title': 'Sample Resource', 'updated_at': '2018-10-11 10:10:15'}]
+        exp_json = [
+            {
+                "active": True,
+                "created_at": "2018-10-10 10:10:10",
+                "id": 1,
+                "title": "Sample Resource",
+                "updated_at": "2018-10-11 10:10:15",
+            }
+        ]
         assert response.status_code == 200
         assert response.json == exp_json
 
     def test_get_resource_by_title(self, client):
         """ Get resource by its title. """
         response = client.get("/resources?title=Sample Resource")
-        exp_json = [{'active': True, 'created_at': '2018-10-10 10:10:10', 'id': 1, 'title': 'Sample Resource', 'updated_at': '2018-10-11 10:10:15'}]
+        exp_json = [
+            {
+                "active": True,
+                "created_at": "2018-10-10 10:10:10",
+                "id": 1,
+                "title": "Sample Resource",
+                "updated_at": "2018-10-11 10:10:15",
+            }
+        ]
         assert response.status_code == 200
         assert response.json == exp_json
 
     def test_add_resource(self, client):
         """ Add resource to database. """
         post_response = client.post(
-            "/resources",
-            json=dict(title='Test Resource', active=True)
+            "/resources", json=dict(title="Test Resource", active=True)
         )
         assert post_response.status_code == 200
-        assert post_response.json['success'] is True
+        assert post_response.json["success"] is True
         get_response = client.get("/resources?title=Test Resource")
-        assert get_response.json[0]['title'] == 'Test Resource'
-        assert get_response.json[0]['id'] == 4
+        assert get_response.json[0]["title"] == "Test Resource"
+        assert get_response.json[0]["id"] == 4
 
     def test_update_resource(self, client):
         """ Update existing resource. """
-        put_response = client.put(
-            "/resources",
-            json=dict(id=4, active=False)
-        )
+        put_response = client.put("/resources", json=dict(id=4, active=False))
         assert put_response.status_code == 200
         get_response = client.get("/resources?id=4")
-        assert get_response.json[0]['active'] is False
+        assert get_response.json[0]["active"] is False
 
     def test_delete_resource(self, client):
         """ Delete existing resource using its ID. """
         delete_response = client.delete("/resources?id=4")
         assert delete_response.status_code == 200
-        assert delete_response.json['success'] is True
+        assert delete_response.json["success"] is True
         get_response = client.get("/resources?id=4")
         assert get_response.json == []
         assert get_response.status_code == 200